refactor: Header 관심사 분리
